refactor(api): headers no longer include other headers
With the exception of the library entry-point (libxnvme.h), then no
other API headers include other headers. That is, all of the
(libxnvme_*.h) do not include any other headers.

It is just the job of libxnvme.h to provide illustrate how to use the
library headers. E.g. which third-party / system-headers are needed by
xnvme. As a consumer of xNVMe, one now have the option of:

A) Default library consumption

  - do: "include <libxnvme.h>"
  - Include **all** of xNVMe definitions
  - Use the third-party definitions as defined by the headers included
    in "libxnvme.h"

B) Tailored library consumption

  - do: "include <libxnvme_COMPONENT.h>"
  - Prefixed with the definitions that you see fit

With the above, then the consumer of the xNVMe library can override
definitions of everything, down to the definitions of errno, uint32_t,
size_t, etc.

* adm: rm unnecessary include
* api/core,dev,opts: redo for modularization
* api/ver: rm headers
* be/internal: rm unnecessary include
* be: rm unnescesary include libxnvme.h
* buf: mv defs. / decl. into libxnvme_buf.h
* buf: rm unnescesary include libxnvme.h
* cmd: mv into libxnvme_cmd.h
* geo: mv include from hdr to src
* geo: rm unnescesary include libxnvme.h
* ident: rm unnescesary include libxnvme.h
* lba: rm unnescesary include libxnvme.h
* opts: mv desf. / decl. into libxnvme_opts.h
* spec: redo include-convention
* spec: rm unnescesary include libxnvme.h
* spec_fs: rm unnescesary include libxnvme.h
* spec_pp: rm unnescesary include libxnvme.h

This change is done in preparation for generating the xNVMe API using
YACE as well as generating ABI-compatible bindings for Python and Rust.
