+++ include/libxnvme_be.h
@@ -8,7 +8,6 @@
 
 #ifndef __LIBXNVME_BE_H
 #define __LIBXNVME_BE_H
-#include <libxnvme.h>
 
 #ifdef __cplusplus
 extern "C" {
