+++ include/libxnvme_dev.h
@@ -8,7 +8,11 @@
 
 #ifndef __LIBXNVME_DEV_H
 #define __LIBXNVME_DEV_H
-#include <libxnvme.h>
+
+enum xnvme_enumerate_action {
+	XNVME_ENUMERATE_DEV_KEEP_OPEN = 0x0, ///< Keep device-handle open after callback returns
+	XNVME_ENUMERATE_DEV_CLOSE     = 0x1  ///< Close device-handle when callback returns
+};
 
 /**
  * Opaque device handle.
@@ -18,6 +22,32 @@
  * @struct xnvme_dev
  */
 struct xnvme_dev;
+
+/**
+ * Signature of callback function used with 'xnvme_enumerate' invoked for each discoved device
+ *
+ * The callback function signals whether the device-handle it receives should by closed, that is,
+ * backend with invoke 'xnvme_dev_close(), after the callback returns or kept open. In the latter
+ * case then it is up to the user to invoke 'xnvme_dev_close()' on the device-handle.
+ *
+ * Each signal is represented by the enum #xnvme_enumerate_action, and the values
+ * XNVME_ENUMERATE_DEV_KEEP_OPEN or XNVME_ENUMERATE_DEV_CLOSE.
+ */
+typedef int (*xnvme_enumerate_cb)(struct xnvme_dev *dev, void *cb_args);
+
+/**
+ * enumerate devices
+ *
+ * @param sys_uri URI of the system to enumerate, when NULL, localhost/PCIe
+ * @param opts Options for instrumenting the runtime during enumeration
+ * @param cb_func Callback function to invoke for each yielded device
+ * @param cb_args Arguments passed to the callback function
+ *
+ * @return On success, 0 is returned. On error, negative `errno` is returned.
+ */
+int
+xnvme_enumerate(const char *sys_uri, struct xnvme_opts *opts, xnvme_enumerate_cb cb_func,
+		void *cb_args);
 
 /**
  * Returns the geometry of the given device
@@ -139,5 +169,25 @@ xnvme_dev_get_be_state(const struct xnvme_dev *dev);
  */
 uint64_t
 xnvme_dev_get_ssw(const struct xnvme_dev *dev);
+
+/**
+ * Creates a device handle (::xnvme_dev) based on the given device-uri
+ *
+ * @param dev_uri File path "/dev/nvme0n1" or "0000:04.01"
+ * @param opts Options for library backend and system-interfaces
+ *
+ * @return On success, a handle to the device. On error, NULL is returned and `errno` set to
+ * indicate the error.
+ */
+struct xnvme_dev *
+xnvme_dev_open(const char *dev_uri, struct xnvme_opts *opts);
+
+/**
+ * Destroy the given device handle (::xnvme_dev)
+ *
+ * @param dev Device handle obtained with xnvme_dev_open()
+ */
+void
+xnvme_dev_close(struct xnvme_dev *dev);
 
 #endif /* __LIBXNVME_NVM */
