+++ include/libxnvme_lba.h
@@ -12,7 +12,6 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-#include <libxnvme.h>
 
 /**
  * Representation of a range of logical-block-addresses aka LBAs
