+++ include/libxnvme_spec_fs.h
@@ -8,10 +8,6 @@
 
 #ifndef __LIBXNVME_SPEC_FS_H
 #define __LIBXNVME_SPEC_FS_H
-#include <assert.h>
-#include <stdint.h>
-#include <sys/types.h>
-#include <libxnvme_util.h>
 
 /**
  * Encapsulate file-system properties as Identify Controller
