+++ include/libxnvme_ver.h
@@ -8,7 +8,6 @@
 
 #ifndef __LIBXNVME_VER_H
 #define __LIBXNVME_VER_H
-#include <libxnvme.h>
 
 /**
  * Produces the "major" version of the library
