+++ include/meson.build
@@ -8,6 +8,7 @@ install_headers('libxnvme.h')
 install_headers('libxnvme_adm.h')
 install_headers('libxnvme_be.h')
 install_headers('libxnvme_buf.h')
+install_headers('libxnvme_cmd.h')
 install_headers('libxnvme_dev.h')
 install_headers('libxnvme_file.h')
 install_headers('libxnvme_geo.h')
@@ -15,7 +16,9 @@ install_headers('libxnvme_ident.h')
 install_headers('libxnvme_lba.h')
 install_headers('libxnvme_libconf.h')
 install_headers('libxnvme_nvm.h')
+install_headers('libxnvme_opts.h')
 install_headers('libxnvme_pp.h')
+install_headers('libxnvme_queue.h')
 install_headers('libxnvme_spec.h')
 install_headers('libxnvme_spec_fs.h')
 install_headers('libxnvme_spec_pp.h')
