+++ include/xnvme_geo.h
@@ -5,9 +5,6 @@
 #ifndef __INTERNAL_XNVME_GEO_H
 #define __INTERNAL_XNVME_GEO_H
 
-#include <libxnvme.h>
-#include <errno.h>
-
 int
 xnvme_geo_yaml(FILE *stream, const struct xnvme_geo *geo, int indent, const char *sep, int head);
